extract header code into own file
